fix(stream): fix encrypt/decrypt of 4GB+ resource
* use explicit uint64 for size and index

+++ modules/streams/include/Tanker/Streams/EncryptionStreamV8.hpp
@@ -32,7 +32,7 @@ public:
 
 private:
   std::optional<std::uint32_t> _paddingStep;
-  std::optional<std::uint32_t> _paddingLeftToAdd;
+  std::optional<std::uint64_t> _paddingLeftToAdd;
 
   tc::cotask<void> encryptChunk();
 };
